% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering_framework.R
\name{loo_encode3}
\alias{loo_encode3}
\title{Do Owen Zhang style leave-one-out encoding of a categorical.
aka likelihood encoding 
https://github.com/DexGroves/hacktoolkit
https://datascience.stackexchange.com/questions/11024/encoding-categorical-variables-using-likelihood-estimation}
\usage{
loo_encode3(id, resp)
}
\arguments{
\item{id}{vector of identifiers to group over}

\item{resp}{vector of response to summarise}
}
\value{
vector of one-left-out summarised response over id
}
\details{
Take the mean of a variable for all rows with the same id except
for the current row, so as to avoid leakage.
}
\examples{
test_data <- data.frame(
  id = c(rep("a", 5), rep("b", 3), rep("c", 2), "d"),
  resp = c(1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 1)
)

loo_encode(test_data$id, test_data$resp)

}
