% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering_framework.R
\name{one_hot_encoding}
\alias{one_hot_encoding}
\title{Framework to encode categorical features 
A compiled list based on HJ van Veen's slide, raddar, and things you will see in kaggle}
\usage{
one_hot_encoding(features = features, mode = "auto")
}
\description{
auto mode will help to remove one column. manual returns n columns with
n variables. You can then choose who is the control.
}
\details{
Objective is to have a one stop framework to deal with categorical variables easy 
Table of contents: 
----------------- 
1) one hot encoding (dummy)
2) Hash encoding (tbd)
3) label encoding 
4) Count encoding 
5) Label count encoding 
6) Leave one out encoding
}
